pyproject.toml
[tool.poetry]
name = "pytest-helm-chart"
version = "0.1.0"
description = "A plugin to provide different types and configs of Kubernetes clusters that can be used for testing."
authors = ["Łukasz Piątkowski <lukasz@giantswarm.io>"]
license="Apache-2.0"
repository="https://github.com/giantswarm/pytest-helm-chart"
readme="README.md"
keywords=["kubernetes","helm","chart"]
include = ["CHANGELOG.md","DCO"]
classifiers=[
    'Development Status :: 4 - Beta',
    'Framework :: Pytest',
    'Intended Audience :: Developers',
    'Topic :: Software Development :: Testing',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: Implementation :: CPython',
    'Operating System :: OS Independent',
    'License :: OSI Approved :: Apache Software License',
]

[tool.poetry.plugins] # Optional super table

[tool.poetry.plugins."pytest11"]
"kube-provider" = "pytest_helm_chart"

[tool.poetry.dependencies]
python = "^3.6"
pytest = "^5.4.2"
pykube-ng = "^20.5.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.2"
tox = "^3.15.1"
tox-pyenv = "^1.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
